+++ EntraAuth/functions/Core/Invoke-EntraRequest.ps1
@@ -53,11 +53,30 @@
 
 	.PARAMETER Raw
 		Do not process the response object and instead return the raw result returned by the API.
+
+	.PARAMETER DeltaSession
+		A hashtable including delta sessions.
+		Use together with the delta endpoints, e.g. for the Graph API's user delta endpoint:
+		https://learn.microsoft.com/en-us/graph/api/user-delta
+		Provide an empty hashtable on the first request, the delta token data will be inserted into it.
+		Provide the same token for subsequent delta requests.
+
+		This allows retrieving changes over time, without having to reload the entire dataset.
+
+	.PARAMETER MinimalDelta
+		When receiving delta data, only return the changed properties (plus a unique identifier), rather than the full object.
+		Only used together with DeltaSession
 	
 	.EXAMPLE
 		PS C:\> Invoke-EntraRequest -Path 'alerts' -RequiredScopes 'Alert.Read'
 	
 		Return a list of defender alerts.
+
+	.EXAMPLE
+		PS C:\> Invoke-EntraRequest -Path 'users/delta' -DeltaSession $delta
+
+		Retrieves all users on first request.
+		Subsequent calls will only return users that have been changed in the meantime.
 #>
 	[CmdletBinding(DefaultParameterSetName = 'default')]
 	param (
@@ -98,7 +117,13 @@
 		$NoPaging,
 
 		[switch]
-		$Raw
+		$Raw,
+
+		[hashtable]
+		$DeltaSession,
+
+		[switch]
+		$MinimalDelta
 	)
 	
 	DynamicParam {
@@ -143,6 +168,7 @@
 			Method = $Method
 			Uri    = Resolve-RequestUri -TokenObject $tokenObject -ServiceObject $serviceObject -BoundParameters $PSBoundParameters
 		}
+		$originalUri = $parameters.Uri
 		
 		if ($PSBoundParameters.Keys -contains 'Body') {
 			if ($Body -is [string]) {
@@ -158,11 +184,16 @@
 			$parameters.Remove('Body')
 		}
 		
-		$parameters.Uri += ConvertTo-QueryString -QueryHash $Query -DefaultQuery $tokenObject.Query
+		$queryClone = $Query.Clone()
+		if ($DeltaSession -and $DeltaSession[$originalUri].Token) {
+			$queryClone['$deltaToken'] = $DeltaSession[$originalUri].Token
+		}
+		$parameters.Uri += ConvertTo-QueryString -QueryHash $queryClone -DefaultQuery $tokenObject.Query
 
 		do {
 			$tempHeader = $tokenObject.GetHeader().Clone() # GetHeader() automatically refreshes expired tokens
 			foreach ($pair in $Header.GetEnumerator()) { $tempHeader[$pair.Key] = $pair.Value }
+			if ($MinimalDelta) { $tempHeader['Prefer'] = 'return=minimal' }
 			$parameters.Headers = $tempHeader
 			Write-Verbose "Executing Request: $($Method) -> $($parameters.Uri)"
 			try { $result = Invoke-RestMethod @parameters -ErrorAction Stop }
@@ -195,6 +226,14 @@
 			if (-not $Raw -and -not $tokenObject.RawOnly -and $result.PSObject.Properties.Where{ $_.Name -eq 'value' }) { $result.Value }
 			else { $result }
 			$parameters.Uri = $result.'@odata.nextLink'
+
+			if ($DeltaSession -and $result.'@odata.deltaLink') {
+				$deltaToken = (($result.'@odata.deltaLink' -split '\?')[1] -split '=')[-1]
+				if (-not $DeltaSession[$originalUri]) {
+					$DeltaSession[$originalUri] = @{}
+				}
+				$DeltaSession[$originalUri]['Token'] = $deltaToken
+			}
 		}
 		while ($parameters.Uri -and -not $NoPaging)
 	}